modules/narrationbuilder/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "narrationbuilder"
version = "0.1.0"
description = "NightChronicles narrationbuilder module: merges narration segments into final narrative"
authors = [{name = "NightChronicles"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "pyyaml>=6.0",
    "rich>=13.7",
    "tenacity>=8.2",
    "httpx>=0.27",
    "openai>=1.17",  # optional at runtime, used when API key present
]

[project.scripts]
narrationbuilder = "narrationbuilder.cli:app"
